Move material class to its own module
