+++ src/compas_cadwork/materials/__init__.py
@@ -1,0 +1,6 @@
+from .material import Material
+
+
+__all__ = [
+    "Material",
+]
