+++ src/compas_cadwork/materials/material.py
